LayoutTests/canvas/philip/tests/2d.composite.uncovered.image.destination-in-expected.txt

< [index] >
2d.composite.uncovered.image.destination-in
drawImage() draws pixels not covered by the source object as (0,0,0,0), and does not leave the pixels unchanged.
References: 2d.composite.destination-in
Actual output:
Expected output:

Passed
